backend/requirements.txt
# --- Web framework ---
fastapi==0.115.0
uvicorn[standard]==0.32.0
python-multipart==0.0.12

# --- Database ---
sqlalchemy==2.0.36
alembic==1.13.3
aiosqlite==0.20.0          # async SQLite driver (dev)
asyncpg==0.30.0            # async PostgreSQL driver (prod)

# --- Validation / settings ---
pydantic==2.9.2
pydantic-settings==2.6.0
email-validator==2.2.0

# --- Auth ---
python-jose[cryptography]==3.3.0
passlib[bcrypt,argon2]==1.7.4
bcrypt==4.0.1
argon2-cffi==23.1.0
pyotp==2.9.0
slowapi==0.1.9
casbin==1.36.3

# --- Real-time ---
python-socketio==5.11.4

# --- ML / inference ---
numpy==1.26.4
pandas==2.2.3
scikit-learn==1.5.2
scipy==1.14.1
xgboost==2.1.2
lightgbm==4.5.0
joblib==1.4.2
tensorflow==2.18.0 ; platform_system != "Darwin" or platform_machine != "arm64"
tensorflow-macos==2.16.2 ; platform_system == "Darwin" and platform_machine == "arm64"

# --- Utilities ---
python-dotenv==1.0.1
loguru==0.7.2
httpx==0.27.2
requests==2.32.3
itsdangerous==2.2.0
prometheus-client==0.21.0

# --- Testing ---
pytest==8.3.3
pytest-asyncio==0.24.0
